requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
psycopg2
requests
pillow
psutil